# Speedup py3k mode