Avoid costly inference in visit_call and visit_attribute
when the attribute name is one that cannot possibly trigger
a message.

This typically leads to a 30-50% speedup when running pylint
in python3 porting mode on most codebases.

Benchmark results on pylint's own code, on a 2018 macbook pro,
with python 3.6.8, using `hyperfine`.

pylint-2.4.2
```
  Time (mean ± σ):     10.071 s ±  1.136 s    [User: 9.724 s, System: 0.246 s]
  Range (min … max):    9.352 s … 12.015 s    5 runs
```

pylint-2.4.2 + this patch
```
  Time (mean ± σ):      6.282 s ±  0.046 s    [User: 6.099 s, System: 0.155 s]
  Range (min … max):    6.232 s …  6.327 s    5 runs
```