@PCManticore 2019-10-30
Hey @huguesb Very nice PR, thank you! I merged this locally to fix the conflict.